Rename ai/algorithms.py to ai/search.py
The module holds one search now rather than a collection of algorithms, so the
name should say so. Import rewrite plus two stale comments: the castling test
explains itself in terms of Board.copy() rather than reaching past it to the
FEN, and the equivalence module no longer describes a comparison that has been
deleted.
